refactor: move config impls to runtime-common
